Eliminate separate error module. Make GetSourceFilesResponse public
